moved loginBox and registerBox to pages
