+++ frontend/src/pages/registerBox.tsx
